requirements.txt
flask==3.1.0
requests==2.32.3
beautifulsoup4==4.13.3
anthropic==0.49.0
python-dotenv==1.1.0
resend==2.26.0
Flask-Limiter==3.8.0
supabase==2.28.3
sentry-sdk==2.57.0
